+++ src/Common/EngineCore/Components/Indexers/RadiantScripting.cs
@@ -1,0 +1,40 @@
+// Sovereign Engine
+// Copyright (c) 2026 opticfluorine
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+using System.Numerics;
+using Sovereign.EngineCore.Components.Types;
+using Sovereign.EngineUtil.Attributes;
+
+namespace Sovereign.EngineCore.Components.Indexers;
+
+/// <summary>
+///     Lua scripting library for evaluating radiant scalar fields.
+/// </summary>
+[ScriptableLibrary("Radiant")]
+public sealed class RadiantScripting(RadiantIndexer radiantIndexer)
+{
+    /// <summary>
+    ///     Evaluates the radiant scalar field of the given category at the queried position.
+    /// </summary>
+    /// <param name="category">Radiant field category.</param>
+    /// <param name="position">Queried position.</param>
+    /// <returns>Sum of the field contributions of all radiant entities in the scanned segments.</returns>
+    [ScriptableFunction("GetValue")]
+    public float GetValue(RadiantCategory category, Vector3 position)
+    {
+        return radiantIndexer.GetValue(category, position);
+    }
+}